Add wheel option

+++ release.py
@@ -159,9 +159,16 @@ def create_option_parser():
         "--pypi",
         action="store_true",
         dest="pypi",
-        help="Initiate a release on PyPi."
+        help="Initiate a release on PyPi. Default is to upload the source distribution. Use the --wheel option to also build a wheel using python-build."
     )
     
+    rel_group.add_option(
+        "--wheel",
+        action="store_true",
+        dest="build_wheel",
+        help="Build a wheel and a source distribution for the PyPi release."
+    )
+
     rel_group.add_option(
         "--cf-hash",
         metavar="HASHFILE",
@@ -349,41 +356,42 @@ def pypi_release(opts, pargs):
     release_dir = pargs[0]
     version = pargs[1]
 
-    # Create temporary release directory
-    tmp_dir = "release_tmp"
-    while (Path(release_dir) / tmp_dir).exists():
-        tmp_dir += "_prime"
-    call(f"mkdir {tmp_dir}", release_dir)
-    
-    # Build tar
-    project = Path(release_dir).name
-    tgz_name = f"{project}-{version}.tar.gz"
-    call(f"tar --exclude=\"./{tmp_dir}\" -zcf \"./{tmp_dir}/{tgz_name}\" . ", release_dir)
+    # Build wheel and source
+    if opts.build_wheel:
+        call("python -m build", release_dir) 
+        # Upload using twine
+        no_tar = True
+        no_whl = True
+        for file in list((release_dir / "dist").iterdir()):
+            if re.search(f".*{version}.*.tar.gz", file.name):
+                no_tar = False
+            if re.search(f".*{version}.*.whl", file.name):
+                no_whl = False
+        if no_tar:
+            call(f"echo \"Warning: No new distribution build. Check for any untracked changes.\"", release_dir)
+        elif no_whl:
+            call(f"echo \"Warning: No wheel found.\"", release_dir)
+        else:
+            call(f"twine upload dist/*{version}*.tar.gz dist/*{version}*.whl", release_dir)
 
-    # IMPORTANT: no wheel built, must be tested on all images~!
-    ## Build distribution (build will fail if there have been no changes since the previous version)
-    # call("python -m build", release_dir) 
-    ## Upload using twine
-    # no_tar = True
-    # no_whl = True
-    # for file in list((release_dir / "dist").iterdir()):
-    #     if re.search(f".*{version}.*.tar.gz", file.name):
-    #         no_tar = False
-    #     if re.search(f".*{version}.*.whl", file.name):
-    #         no_whl = False
-    # if no_tar:
-    #     call(f"echo \"Warning: No new distribution build. Check for any untracked changes.\"", release_dir)
-    # elif no_whl:
-    #     call(f"echo \"Warning: No wheel found.\"", release_dir)
-    # else:
-    #     call(f"twine upload dist/*{version}*.tar.gz dist/*{version}*.whl", release_dir)
-
-    # Upload using twine
-    # Only source is uploaded
-    call(f"twine upload \"./{tmp_dir}/{tgz_name}\"", release_dir)
+    # Only upload source
+    else:
+        # Create temporary release directory
+        tmp_dir = "release_tmp"
+        while (Path(release_dir) / tmp_dir).exists():
+            tmp_dir += "_prime"
+        call(f"mkdir {tmp_dir}", release_dir)
     
-    # Cleanup
-    call(f"rm -rf {tmp_dir}", release_dir)
+        # Build tar
+        project = Path(release_dir).name
+        tgz_name = f"{project}-{version}.tar.gz"
+        call(f"tar --exclude=\"./{tmp_dir}\" -zcf \"./{tmp_dir}/{tgz_name}\" . ", release_dir)
+
+        # Upload using twine
+        call(f"twine upload \"./{tmp_dir}/{tgz_name}\"", release_dir)
+    
+        # Cleanup
+        call(f"rm -rf {tmp_dir}", release_dir)
 
 
 # Generate SHA256 Hash for a Conda-Forge Release
